refactor: batch SpoolmanDB downloads and share CommunityCatalogItem type
